refactor: Discord VoiceClient와 관련된 모듈 리팩토링
